# python-flask-server/swagger_server/models/login.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from swagger_server.models.base_model_ import Model
from swagger_server import util


class Login(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """

    def __init__(self, name: str=None, phone: str=None):  # noqa: E501
        """Login - a model defined in Swagger

        :param name: The name of this Login.  # noqa: E501
        :type name: str
        :param phone: The phone of this Login.  # noqa: E501
        :type phone: str
        """
        self.swagger_types = {
            'name': str,
            'phone': str
        }

        self.attribute_map = {
            'name': 'name',
            'phone': 'phone'
        }

        self._name = name
        self._phone = phone

    @classmethod
    def from_dict(cls, dikt) -> 'Login':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The Login of this Login.  # noqa: E501
        :rtype: Login
        """
        return util.deserialize_model(dikt, cls)

    @property
    def name(self) -> str:
        """Gets the name of this Login.

        notice type  # noqa: E501

        :return: The name of this Login.
        :rtype: str
        """
        return self._name

    @name.setter
    def name(self, name: str):
        """Sets the name of this Login.

        notice type  # noqa: E501

        :param name: The name of this Login.
        :type name: str
        """

        self._name = name

    @property
    def phone(self) -> str:
        """Gets the phone of this Login.

        notice content  # noqa: E501

        :return: The phone of this Login.
        :rtype: str
        """
        return self._phone

    @phone.setter
    def phone(self, phone: str):
        """Sets the phone of this Login.

        notice content  # noqa: E501

        :param phone: The phone of this Login.
        :type phone: str
        """

        self._phone = phone