# python-flask-server/swagger_server/swagger/swagger.yaml
...
swagger: "2.0"
info:
  description: "Chibi power face recognition system"
  version: "1"
  title: "Yc"
host: "yc.api.ink"
basePath: "/v1"
schemes:
- "http"
consumes:
- "application/json"
produces:
- "application/json"
paths:
  /phone/login:
    post:
      tags:
      - "phone"
      summary: "phone Product warehousing"
      description: "phone Product warehousing"
      operationId: "phone_login_post"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "data"
        required: true
        schema:
          $ref: "#/definitions/Login"
      responses:
        200:
          description: "return status"
          schema:
            $ref: "#/definitions/success"
        500:
          description: "return status"
          schema:
            $ref: "#/definitions/success"
      x-swagger-router-controller: "swagger_server.controllers.phone_controller"
  /phone/outlist:
    post:
      tags:
      - "phone"
      summary: "phone Product warehousing"
      description: "phone Product warehousing"
      operationId: "phone_outlist_post"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "data"
        required: true
        schema:
          $ref: "#/definitions/outlist"
      responses:
        200:
          description: "return status"
          schema:
            $ref: "#/definitions/success"
        500:
          description: "return status"
          schema:
            $ref: "#/definitions/success"
      x-swagger-router-controller: "swagger_server.controllers.phone_controller"
  /phone/outcommit:
    post:
      tags:
      - "phone"
      summary: "phone Product warehousing"
      description: "phone Product warehousing"
      operationId: "phone_outcommit_post"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "data"
        required: true
        schema:
          $ref: "#/definitions/outcommit"
      responses:
        200:
          description: "return status"
          schema:
            $ref: "#/definitions/success"
        500:
          description: "return status"
          schema:
            $ref: "#/definitions/success"
      x-swagger-router-controller: "swagger_server.controllers.phone_controller"
  /phone/storage:
    post:
      tags:
      - "phone"
      summary: "phone Product warehousing"
      description: "phone Product warehousing"
      operationId: "phone_storage_post"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "data"
        required: true
        schema:
          $ref: "#/definitions/storageData"
      responses:
        200:
          description: "return status"
          schema:
            $ref: "#/definitions/success"
        500:
          description: "return status"
          schema:
            $ref: "#/definitions/success"
      x-swagger-router-controller: "swagger_server.controllers.phone_controller"
  /phone/out:
    post:
      tags:
      - "phone"
      summary: "phone Product warehousing"
      description: "phone Product warehousing"
      operationId: "phone_out_post"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "data"
        required: true
        schema:
          $ref: "#/definitions/OutData"
      responses:
        200:
          description: "return status"
          schema:
            $ref: "#/definitions/success"
        500:
          description: "return status"
          schema:
            $ref: "#/definitions/success"
      x-swagger-router-controller: "swagger_server.controllers.phone_controller"
  /phone/use:
    post:
      tags:
      - "phone"
      summary: "phone use post"
      description: "phone use post"
      operationId: "phone_use_post"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "data"
        required: true
        schema:
          $ref: "#/definitions/phoneUseLists"
      responses:
        200:
          description: "return status"
          schema:
            $ref: "#/definitions/success"
        500:
          description: "return status"
          schema:
            $ref: "#/definitions/success"
      x-swagger-router-controller: "swagger_server.controllers.phone_controller"
  /phone/maintain:
    post:
      tags:
      - "phone"
      summary: "phone maintain post"
      description: "phone maintain post"
      operationId: "phone_maintain_post"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "data"
        required: true
        schema:
          $ref: "#/definitions/phoneMaintainLists"
      responses:
        200:
          description: "return status"
          schema:
            $ref: "#/definitions/success"
        500:
          description: "return status"
          schema:
            $ref: "#/definitions/success"
      x-swagger-router-controller: "swagger_server.controllers.phone_controller"
  /phone/token:
    post:
      tags:
      - "phone"
      summary: "phone maintain post"
      description: "phone token post"
      operationId: "phone_token_post"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "data"
        required: true
        schema:
          $ref: "#/definitions/phonetoken"
      responses:
        200:
          description: "return status"
          schema:
            $ref: "#/definitions/phonetokens"
        500:
          description: "return status"
          schema:
            $ref: "#/definitions/success"
      x-swagger-router-controller: "swagger_server.controllers.phone_controller"
...
definitions:
...
  OutData:
    properties:
      work:
        type: "string"
        description: "Workstation"
      type:
        type: "string"
        description: "terminal type"
      model:
        type: "string"
        description: "Terminal model"
      people:
        type: "string"
        description: "Installation staff"
      code:
        type: "string"
        description: "Bar code"
    example:
      code: "code"
      work: "work"
      model: "model"
      type: "type"
      people: "people"
...
  phoneMaintainLists:
    properties:
      work:
        type: "string"
        description: "workstation"
      order:
        type: "string"
        description: "Complaint number"
      type:
        type: "string"
        description: "Replace terminal type"
      ocode:
        type: "string"
        description: "Original barcode"
      code:
        type: "string"
        description: "new barcode"
      personnel:
        type: "string"
        description: "Installation staff"
    example:
      code: "code"
      work: "work"
      ocode: "ocode"
      personnel: "personnel"
      type: "type"
      order: "order"
...
  phoneUseLists:
    properties:
      work:
        type: "string"
        description: "workstation"
      business:
        type: "string"
        description: "business type"
      order:
        type: "string"
        description: "Work order"
      modem:
        type: "string"
        description: "Modem bar code"
      gateway:
        type: "string"
        description: "Gateway barcode"
      box:
        type: "string"
        description: "Set-top box barcode"
      hemu:
        type: "string"
        description: "hemu bar code"
      phone:
        type: "string"
        description: "Fixed phone barcode"
      personnel:
        type: "string"
        description: "Installation staff"
    example:
      business: "business"
      hemu: "hemu"
      phone: "phone"
      work: "work"
      modem: "modem"
      box: "box"
      personnel: "personnel"
      gateway: "gateway"
      order: "order"
...
  storageData:
    type: "object"
    properties:
      work:
        type: "string"
        description: "workstation"
      type:
        type: "string"
        description: "terminal type"
      factory:
        type: "string"
        description: "factory"
      model:
        type: "string"
        description: "model"
      principal:
        type: "string"
        description: "Warehousing"
      code:
        type: "string"
        description: "Bar code"
...
  success:
    properties:
      code:
        type: "integer"
        description: "system return code"
      msg:
        type: "string"
        description: "system return news"
    example:
      msg: "msg"
      code: 0
...
  phonetoken:
    properties:
      url:
        type: "string"
        description: "url"
    example:
      url: "www.baidu.com"
  outlist:
    properties:
      outman:
        type: "string"
        description: "the pepole for outbound"
  outcommit:
    properties:
      people:
        type: "string"
        description: "the pepole for outbound"
      type:
        type: "string"
        description: "the pepole for outbound"
      count:
        type: "string"
        description: "the pepole for outbound"
      model:
        type: "string"
        description: "the pepole for outbound"
  phonetokens:
    properties:
      code:
        type: "integer"
        description: "system return code"
      msg:
        type: "string"
        description: "system return news"
      result:
        type: "array"
        items:
          $ref: "#/definitions/phonetokenss"
    example:
      msg: "msg"
      code: 0
      result:
      - code: "code"
        times: "times"
        noncestr: "noncestr"
  phonetokenss:
    properties:
      code:
        type: "string"
        description: "system return code"
      times:
        type: "integer"
        description: "system return news"
      noncestr:
        type: "string"
        description: "system return news"
...
  Login:
    properties:
      name:
        type: "string"
        description: "notice type"
      phone:
        type: "string"
        description: "notice content"
    example:
      name: "name"
      phone: "phone"